Input: byd - initial implementation of the BYD PS/2 touch pad protocol
This adds multitouch support to the BYD Touch Pad model BTP10463
by using the vendor specific initialisation sequence that makes the
touchpad firmware report additional codes representing gestures and
different number of fingers.

However, like the vendor's Windows driver, these are not configurable
and it is unknown whether this touchpad supports absolute mode. The
driver exposes the vendor-specific codes as different button-presses
which can be then mapped by the user as evdev supports up to 16 button
keys.  Since the pinch works in the same manner as the scroll, it is
also exposed as a scroll as REL_MISC.